requirements.txt
torch
torchvision
gradio
numpy
pillow
scikit-learn
matplotlib
